# artifacts/api-server/src/routes/notifications.ts
import { Router } from "express";
import { db } from "../lib/db";
import { notifications } from "@workspace/db/schema";
import { eq, and, isNull, desc } from "drizzle-orm";
import { requireAuth, AuthRequest } from "../middleware/auth";
import { notificationsEmitter } from "../lib/notifications-emitter";

const router = Router();

router.get("/notifications/stream", requireAuth, (req: AuthRequest, res) => {
  res.setHeader("Content-Type", "text/event-stream");
  res.setHeader("Cache-Control", "no-cache, no-transform");
  res.setHeader("Connection", "keep-alive");
  res.setHeader("X-Accel-Buffering", "no");
  res.flushHeaders?.();

  res.write(`event: ready\ndata: {"ok":true}\n\n`);

  const send = (event: string, data: unknown) => {
    res.write(`event: ${event}\n`);
    res.write(`data: ${JSON.stringify(data)}\n\n`);
  };

  const unsubscribe = notificationsEmitter.subscribeForUser(req.userId!, (ev) => {
    send("notification", ev.notification);
  });

  const heartbeat = setInterval(() => {
    res.write(`: heartbeat ${Date.now()}\n\n`);
  }, 25_000);

  req.on("close", () => {
    clearInterval(heartbeat);
    unsubscribe();
  });
});

router.get("/notifications", requireAuth, async (req: AuthRequest, res) => {
  const unreadOnly = req.query["unreadOnly"] === "true";

  const rows = await db
    .select()
    .from(notifications)
    .where(
      unreadOnly
        ? and(eq(notifications.userId, req.userId!), isNull(notifications.readAt))
        : eq(notifications.userId, req.userId!)
    )
    .orderBy(desc(notifications.createdAt))
    .limit(50);

  res.json({ notifications: rows });
});

router.patch("/notifications/:id/read", requireAuth, async (req: AuthRequest, res) => {
  const id = Number(req.params["id"]);

  const [updated] = await db
    .update(notifications)
    .set({ readAt: new Date() })
    .where(
      and(eq(notifications.id, id), eq(notifications.userId, req.userId!))
    )
    .returning();

  if (!updated) {
    res.status(404).json({ error: "Notifikasi tidak ditemukan" });
    return;
  }

  res.json({ message: "Notifikasi ditandai telah dibaca" });
});

router.patch("/notifications/read-all", requireAuth, async (req: AuthRequest, res) => {
  await db
    .update(notifications)
    .set({ readAt: new Date() })
    .where(
      and(
        eq(notifications.userId, req.userId!),
        isNull(notifications.readAt)
      )
    );

  res.json({ message: "Semua notifikasi ditandai telah dibaca" });
});

export default router;