# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Trading Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password/question:
    post:
      operationId: getForgotPasswordQuestion
      tags: [auth]
      summary: Get security question for email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordQuestionBody"
      responses:
        "200":
          description: Security question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityQuestionResponse"
  /auth/forgot-password/verify:
    post:
      operationId: verifySecurityAnswer
      tags: [auth]
      summary: Verify security answer and get reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifySecurityAnswerBody"
      responses:
        "200":
          description: Answer verified, reset token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetTokenResponse"
        "401":
          description: Wrong answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password/reset:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password:
    patch:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/security-question:
    patch:
      operationId: changeSecurityQuestion
      tags: [auth]
      summary: Change security question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeSecurityQuestionBody"
      responses:
        "200":
          description: Security question updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        selectedMode:
          type: string
          enum: [beginner, pro]
        themePreference:
          type: string
          enum: [light, dark]
        securityQuestion:
          type: string
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - role
        - selectedMode
        - themePreference
        - onboardingCompleted
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
        selectedMode:
          type: string
          enum: [beginner, pro]
          default: beginner
        securityQuestion:
          type: string
        securityAnswer:
          type: string
        rememberMe:
          type: boolean
          default: false
      required:
        - email
        - password
        - displayName
        - securityQuestion
        - securityAnswer
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
      required:
        - email
        - password
    ForgotPasswordQuestionBody:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    SecurityQuestionResponse:
      type: object
      properties:
        securityQuestion:
          type: string
        email:
          type: string
      required:
        - securityQuestion
        - email
    VerifySecurityAnswerBody:
      type: object
      properties:
        email:
          type: string
          format: email
        securityAnswer:
          type: string
      required:
        - email
        - securityAnswer
    ResetTokenResponse:
      type: object
      properties:
        resetToken:
          type: string
        message:
          type: string
      required:
        - resetToken
        - message
    ResetPasswordBody:
      type: object
      properties:
        resetToken:
          type: string
        newPassword:
          type: string
          minLength: 6
      required:
        - resetToken
        - newPassword
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        selectedMode:
          type: string
          enum: [beginner, pro]
        themePreference:
          type: string
          enum: [light, dark]
        onboardingCompleted:
          type: boolean
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required:
        - currentPassword
        - newPassword
    ChangeSecurityQuestionBody:
      type: object
      properties:
        currentPassword:
          type: string
        securityQuestion:
          type: string
        securityAnswer:
          type: string
      required:
        - currentPassword
        - securityQuestion
        - securityAnswer
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        targetRole:
          type: string
          nullable: true
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, error]
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - type
        - createdAt
    NotificationsList:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required:
        - notifications
        - unreadCount
...